+++ plugins/discourse-narrative-bot/app/jobs/regular/bot_input.rb
