+++ plugins/discourse-narrative-bot/app/jobs/regular/narrative_init.rb
