+++ plugins/discourse-narrative-bot/app/jobs/regular/send_default_welcome_message.rb
